ProgramacionIVContenido/tareas/tarea_6_Transformacion_de_Aplicacion_CLI_a_Aplicacion_Web_con_Flask/requirements.txt
flask==3.0.3
redis==5.0.7
python-dotenv==1.0.1
